solved some server problems , changed routes of some page , fixed Sign in and  Sign up buttons, now when you sign in with existing account you will be directed to dashboard webpage
