+++ Plans/plans.html
@@ -15,12 +15,12 @@
         <div class="left-section">
             <img src="../img/meteciclelogo.png" alt="Metecicle Logo">
             <ul>
-                <li><a href="../HomePage/index.html">Home</a></li>
-                <li><a class="active" href="plans.html">Pricing</a></li>
-                <li><a href="about/about.html">About</a></li>
+                <li><a href="/">Home</a></li>
+                <li><a class="active" href="/Pricing">Pricing</a></li>
+                <li><a href="/about">About</a></li>
                 <li><a href="">Support</a></li>
                 <li><a href="">Contact</a></li>
-                <li><a href="">More:</a></li>
+                <li><a href="/DashBoard">DashBoard</a></li>
             </ul>
         </div>
     </nav>
