+++ Server/.env
@@ -1,2 +1,2 @@
-MONGODBURI = "enter yours "
+MONGODBURI = ""
 PORT=3000
