+++ Server/server.js
@@ -4,30 +4,33 @@ require('dotenv').config();
 const cors = require('cors');
 const path = require('path');
 const rateLimit = require('express-rate-limit');
-const helmet = require('helmet');
 const routes = require('./route.js');
-
+const staticFiles = require('./staticFiles.js'); // Now an array of functions
 const app = express();
+const crypto = require('crypto');
 const URI = process.env.MONGODBURI;
 const port = process.env.PORT || 3000;
 
-// Security Middleware
-app.use(helmet());
-app.use(cors());
-app.use(express.json());
-
-// Rate Limiting
 const limiter = rateLimit({
     windowMs: 15 * 60 * 1000, // 15 minutes
     max: 100, // Limit each IP to 100 requests per windowMs
 });
 app.use('/data', limiter);
 
+// Security Middleware (without helmet)
+app.use((req, res, next) => {
+    res.locals.nonce = crypto.randomBytes(16).toString('base64'); // Generate nonce
+    next();
+});
+
+app.use(cors());
+app.use(express.json());
+
 // Serve static files
-app.use(express.static(path.join(__dirname, '../')));
+staticFiles.forEach((setupStaticFile) => setupStaticFile(app));
 
 // MongoDB Connection
-mongoose.connect(URI)
+mongoose.connect(process.env.MONGODBURI)
     .then(() => console.log('MongoDB connection established'))
     .catch((err) => {
         console.error('MongoDB connection error:', err);
