+++ Sign-In/sign-in.html
@@ -19,7 +19,6 @@
             <span id="error-message"></span>  <!-- To display error message -->
         </form>
             
-        <p class="error-message"></p>
     </div>
     <script src="sign-in.js"></script>
 </body>
