+++ about/about_styles.css
@@ -1,87 +1,6 @@
-/* General Body */
-body {
-    background-image: url('img/logotest1.png');
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center;
-    margin: 0;
-    font-family: Arial, sans-serif;
-}
-
-/* Navigation Bar */
-nav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 15px 20px;
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-    position: sticky;
-    top: 0;
-    z-index: 1000;
-}
-
-.left-section {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-}
-
-.left-section img {
-    max-height: 50px;
-}
-
-.left-section ul {
-    display: flex;
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    gap: 15px;
-}
-
-.left-section ul li a {
-    text-decoration: none;
-    font-size: 16px;
-    color: #333;
-    transition: color 0.3s ease-in-out;
-}
-
-.left-section ul li a:hover {
-    color: #4caf50;
-    text-decoration: underline;
-}
-.right-section {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.right-section button {
-    padding: 8px 16px;
-    border: 2px solid transparent;
-    background-color: #fcb69f;
-    color: #fff;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 14px;
-    transition: transform 0.3s, background-color 0.3s, border-color 0.3s;
-}
-
-.right-section button.signup {
-    background-color: #4caf50;
-}
-
-.right-section button:hover {
-    transform: scale(1.1);
-    background-color: #333;
-    color: #fff;
-    border-color: #fcb69f;
-}
-
 /* About Header */
 .about-header {
     text-align: center;
-    margin-top: 20px;
     padding: 60px 20px;
     background: linear-gradient(135deg, #ffecd2 0%, #fcb69f 100%);
     border-radius: 10px;
